12/12 イベント一覧表示（条件追加）、セッションテーブル更新機能追加

+++ logout.php
@@ -14,5 +14,9 @@ $loader = new \Twig\Loader\FilesystemLoader(Bootstrap::TEMPLATE_DIR);
 $twig = new \Twig\Environment($loader,[
     'cache' => Bootstrap::CACHE_DIR
 ]);
-$ses->destroy();
-var_dump($_SESSION);
+$ses->checkSession();
+
+if(isset($_SESSION['member_id'])){
+    $ses->deleteSession();
+    $ses->destroy();
+}
